Add *move-to-branch variants
There're already *uncommit-to-branch, which prompts for a commit in the according range. But sometimes one wants to move the entire range to a new branch, and the query is tedious and blocks automation.
So add a variant that uses "move" instead of "uncommit".
Note: We need the -- separation in several variants, as the -b|--branch argument taken by uncommit-to-branch is also interpreted by various rev-range-* wrappers themselves.