BUG: *uncommit-to-branch cannot take -b|--branch argument
(Or any other user-passed argument.)
git-selectedcommit-command expects the <range> directly after the GIT-COMMAND. Unfortunately, there's already a lot of parsing going on there, and the <range> itself has such a loose syntax that it's difficult to discern.
Add a special --range-is-last flag catered to the rev-range-* behavior of appending the range to any alias and user-provided arguments.