git-prune: CHG: Rename to rprune, take <remote> as first argument, turn wrapper scripts into aliases
rprune is more consistent with other commands, and avoids problems from overriding the built-in "git prune" (which I wasn't aware of, as is a rather technical low-level tool anyway).
By taking the <remote> as the first argument, we're consistent with the other git-r* commands. As the [ou]prune aliases are meant to be invoked by the user, this odd argument ordering doesn't matter.