git-archivebr: ENH: Also support archiving of remote branches
It can be useful to archive the (untouched) origin branch after the local branch has already been rebased onto an updated upstream.
The remote is dropped from the archived tag name. Archived branches are sorted into localBranches and originBranches, and only added when successfully archived. These arrays are then used for the cleanup.