ENH: Allow pruning of remote branches via -R|--reverse
I thought about adding a --except flag to git-odeletebr, but the specification of branches not to delete is difficult and error-prone. Instead (at least in my current situation), the local branch state is correct and we want that to be reflected in the remote as well, so this works like a reverse pruning, and this command is ideal to extend.