WIP: Convert subcobr alias to separate script
In order to greatly enhance the functionality and adapt it to my use cases.

I think there's also a bug in the original (that I got from Stack Overflow): The --recursive doesn't retrieve remote branches of nested submodules, because the $toplevel looks in the superproject's .gitmodules, but submodules have their own .gitmodules.