git-logdelta: ENH: Allow two-stage invocation with passed record
If the client needs to control the working copy update themselves, the command can now be invoked twice. On the first invocation, it takes a snapshot of the current working copy (and submodule checkouts if --submodule is given). On the second invocation, it receives that snapshot and compares it with the current state.