CHG: Default to showing a status, add --status and --all to control this
A separate command to show the current status of the submodules would duplicate a lot of the functionality.
Also, the immediate checkout of _all_ submodules if no arguments are given is very dangerous.
Default to showing the status, unless at least one SUBMODULE is given. Status of individual SUBMODULE(s) can still be obtained via --status.