subup: Refine and add sub{co,fetchup} variants
There's a lot of complexity in "git submodule update": It can be used to do a bulk-update of the superproject with newer changes in subprojects via --remote, and the fetch can be included or not. According to my notation, that's "fetchup" and "up", resp.
If you don't want the bulk update, inside each submodule "git up" can be used as well.
Without --remote, "git submodule update" works the other way around: It pushes the submodule state recorded in the superproject onto all submodules, i.e. functions like "git checkout" limited to submodules.
The --rebase influences the behavior when a submodule is not a detached checkout, but instead has a branch checked out. Any local changes will then be rebased (or merged with --merge), instead of doing a detached checkout. If the submodule is detached, the behavior is identical; however, the messages are a bit different:

> $ git submodule update --init --recursive
> Submodule path 'SUBMODULE': checked out 'e504c3be49c581f3a1518c555f29f4a06d5524ab'
>
> $ git submodule update --init --rebase --recursive
> First, rewinding head to replay your work on top of it...
> Fast-forwarded HEAD to e504c3be49c581f3a1518c555f29f4a06d5524ab.
> Submodule path 'SUBMODULE': rebased into 'e504c3be49c581f3a1518c555f29f4a06d5524ab'