ENH: Allow reflog filtering of only commits with --commits-only
There are two main targets in the reflog: Usually we want to restore to a certain (lost) commit, but sometimes we've done a lot of resetting, rebasing etc. and want to return to a good state.
The command currently is optimized for the latter, but the former is also very important, if not used more frequently. Therefore, add a different mode where only commits (amends, which are listed as "commit (amend):") are offered.
The parsing is a bit hairy, as we first need to pick apart the command-line arguments to detect the custom flag, pass it inside GIT_SELECTEDCOMMIT_COMMAND, then parse it again in the recursive invocation to set up the filter (for which we need to turn off the ref name decoration (%d)).