BUG: AWK parsing cannot detect submodule.SUBNAME.branch = .
Because we define . as a separator in order to easily pick a part the submodule.SUBNAME.branch (but that's wrong and would break with a submodule name containing a ".").
Only use "=" as separator (still breaking on submodule names containing that, but less likely now), and match via regexp.
For the branch config, the submodule name needs to be extracted from the key.