CHG: Reset an existing local branch to the remote branch instead of just checking out the corresponding local branch
If we just want to switch to the last checked-out state of a remote branch, we can simply "git checkout LOCAL-BRANCH". The DWIM action is to automatically reset the local branch to the remote one (assuming changes have been fetched in the meantime); this avoids a checkout of an outdated state followed by an explicit update (so potentially larger working copy changes if stuff got reorganized).