Explicitly pass -b $branch when checking out a new local branch
Though --track has some heuristic to determine the local branch name, the documentation sounds as if it wouldn't correctly deal with branches that contain slashes, and in general could fail if the "guessing" goes wrong. As we've already determined the desired local branch name according to our algorithm, pass it to be safe.