ENH: Allow -- before WORKING-COPY to end parsing of those, but keep parsing command options (until a second -- [SIMPLECOMMAND])
The problem in the git-wcs-in-dir-do-command wrappers (like Unixhome-ingo, swdev) is that we don't want to force the user to pass the -- before SIMPLECOMMAND, but also need to take options arguments like --no-git-color.
For the former, we need to use -- "$@"; however, that precludes the use of further options, as -- meant end of parsing, and what follows is SIMPLECOMMAND, period. If we drop the -- from the wrapper, however, the mandatory separator before SIMPLECOMMAND has to be supplied by the user.
Change the parsing so that a first -- just sets an isDoneWorkingCopies flag, and that will then break out of the parsing if a non-option is encountered, instead of taking the argument as a WORKING-COPY. A second -- will still immediately exit parsing.