git-sub[iI]: BUG: Reimplementation: Parse git status porcelain version 2
With the normal "git status" output, despite the --ignore-submodules configuration, it's not possible to clearly differentiate between changes in the superproject's state of the submodule and the submodule's state.
Fortunately, the porcelain version 2 output has a dedicated field that not only allows us to detect submodules (obviating git-subfilter), but do simple filtering by matching a character at a fixed position.
With --no-renames, we don't even need to consider any renamings of submodules (if that's actually possible?)