Robustness: Ensure that we have access to terminal before mutating the working copy
Usually, commands like git-reintegrate abort if necessary input cannot be obtained. But some wrappers already do mutations on their own before delegating (here: to git-reintegrate), and that would leave the working copy in an unexpected state when the actual action cannot be executed.
Check that we can interactively query the user (i.e. writing to the terminal and obtaining input from it).