ENH: Detect mismatch with tracked upstream branch and try fast-forwarding first
After working on a feature branch, we likely have fetched upstream but possibly have not fast-forwarded our local master branch (as we've been working exclusively on that branch). Now when we simply reintegrate, ignorant of that situation, we may then have to merge with upstream after the push failed.
It's better to first update our local branch to upstream (which likely is a fast-forward, as long as we haven't done any other local development in parallel), and only then merge in our changes, as that's just a single merge instead of two.