git-rev-range-{incom,outgo}ing: FIX: Use passed -b <branch> both for remote and local
As documented; it doesn't make that much sense to compare a passed branch to the current branch's remote (and that would (for master) then have to be called "om{out,in}*").
As we now need to parse -b|--branch ourselves, the functionality of git-branch-command is not required any longer here.