git-reintegrate: ENH: Usability: Check whether fast-forwarding to tracked branch is actually possible and ask accordingly
I had just unmerged a branch, added some more commits, and then wanted to reintegrate with master when the command decided to fast-forward origin and mess up the merge. Granted, I could have force-pushed origin first, but I like to have the original state as a backup (less messy to reset to the origin branch than to deal with the local reflog). Now it would ask me for confirmation and I could politely decline.