CHG: Make git-fetchdate store per-remote timestamps and rewrite git-syncdate to consider per-remote fetchdates
Store the timestamps in a custom database in the Git dir (.git/fetches), triggered via the undocumented --set REMOTE [...] command-line option.
git-syncdate now needs to consider fetches and pushes separately to determine the latest one, and then also do a unification of fetch(es) and push(es) that happened close together. But this basically is just some shifting around of the original algorithm.