Adapt: progressNotificationToAddendum has been incorporated into progressNotification as an extension
This way, any sink is handled uniformly, and we can get rid of the ugly special case.
Don't copy the progressNotification sinks any longer; simply refer to that command. That additional man lookup is a bit less convenient for the user, but greatly reduces the maintenance effort, which is important since the functionality has already metastasized into 3 clones.