git-up: ENH: Fast-forward a corresponding branch of a detached checkout if it can be checked out
With my "./superproject/submodule along ./submodule" checkout, I have the problem that master usually isn't checked out (only when I'm currently reintegrating a submodule), so my local branch falls further and further behind (unless I'm doing a reintegration).
Add logic to fast-forward a detached checkout of a local branch, but only if the working copy is clean and a checkout is possible.
With that, my "all dev"-update will attempt this both within the submodule and in the separate worktree, and one of those should succeed.