git-reintegrate: Also delete origin branch of the same name when it's not tracked
When working with submodules, a branch may have been pushed because a superproject commit references it, but that does not establish a tracking relationship.
We still would like a cleanup of that branch (as long as it's in our GitHub clone). This can be achieved by replacing "git deletebr" with "git-rldeletebr origin"; the --ignore-missing-in-origin suppresses any error should the origin branch not exist.
Note: git-rldeletebr also does a tracked remote branch lookup, but will only accept one belonging to the passed REMOTE. As we may reintegrate an upstream (or whatever) branch, we cannot simply delegate once to git-rldeletebr, but still need to do the tracking branch query ourselves.