Support --[no-]color (first) arg as a no-op
So that when one of my custom commands is used as GIT-SIMPLECOMMAND in a git-*wcdo command (which captures stdout for its header / pager and therefore needs special handling of the color), we don't need to pass the --no-git-color flag to it. Forgetting that (and then getting a list of Git errors about --color=always) is a frequent source of errors.