git-whowhen: CHG: Default to authors and replace --authored with --committer
As the authors are more important (e.g. a frequent committer may be "GitHub"), and this is then aligned with "git whodid".
I took the --authored argument from git-datediff, where it's the other way around: We're rather interested in commit activities, not the actual authoring.
The --committer argument is inspired from git-shortlog.