git-inout: Use the detached effective branch name in the output
The commit ID is confusing here; better state our (assumed) sources.
A complication here is that baseRev is used for a sanity check, so we must not embellish this one. Establish a separate baseName (and otherName for consistency, although that one's currently not assigned to) and prefer that when outputting, falling back to baseRev.