git-parse-last-remotebr-argument: Fall back to a detached effective branch if last branch argument given
So that aliases like od and olog work like oinout.
It might be more correct to take a --detached-effective-branch argument as some [uo]* aliases (omerge?) might not make sense with an "effective" branch. However, I think the error would be clear and it's too much effort to go through all aliases and add that flag.