git-subdo-core: ENH: Allow suppression of "No submodules found!" error and custom exit status via GIT_SUBDO_EXIT_STATUS_NO_SUBMODULES
Instead of a -q|--quiet flag. If the user passes a custom exit status, they are able to recognize the problem, and could do the error reporting themselves (or may want to suppress the error).
Change the default to 99 (no data), as that makes more sense than 1 and can be better distinguished from COMMAND errors.