git-wips: Adapt: Use new GIT_SUBDO_EXIT_STATUS_NO_SUBMODULES to avoid a preceding git-hassubmodules check
When there are no submodules, the "git submodule foreach" iteration that git-subdo-core runs (with recursive invocation of itself) is then a no-op and not less efficient; git-hassubmodules invokes "git sublist" which does a "git submodule foreach", too.
The environment variable also suppresses the "No submodules found!" error message that is not desired for wips, too.