git-{brref,localbr,sub,wc}do[-core]: ENH: Allow treating COMMAND returning 99 as no-op instead of error via --status-99-is-noop
This is useful for running checks on all subjects, as it allows to differentiate between success (0), no action necessary (99), failure (1-98, 100+).
The logic is delegated to my status99Reduction helper.
Also exit with 99 if no subjects ([local] branches, submodules) found (but for git-wcdo, this is an invocation error).

Bonus: Consistently run the COMMAND(s) in a subshell; this not just protects against --command 'exit 42', but also prevents accidental messing with internal variables.