git-commitidentical: ENH: Allow referencing a commit in a different working copy
This is super cool for identical changes across submodules, because we can use the superproject / first submodule's commit and apply that to all (dirty) submodule repositories.
The syntax is similar to git-inside, so we can simply take the "git inside lo1 SHA -- PATH" from the git commit wrapper, and turn it into "git commitidentical SHA -- PATH".
git-commitlike inherits this new ability automatically; we still need to separately extend git-correct to get the same for git-commitrelate.