ENH: Keep user/repo@ID link to last submodule commit when no PR and --qualified-with-userrepo is given
Ideally, we'd be linking to all commit IDs of the submodule commits, but from "git diff --submodule=log" we'll only get the FROM..TO. I'd have to do a "git submodule foreach git log FROM..TO" to obtain that.
Let's keep it at that; there's likely a pull request if there are many submodule commits, anyway, and in GitHub it's also easy to navigate the prior commits (only problem is knowing where to stop).