BUG: Parsing of fileAgeArgs assumes a trailing REMOTE argument
But that is optional! Would need to check for >= 2 instead of >= 3, but let's actually parse all the age options (like in git-stalebr) instead of assuming it's one.
As there are several different arguments (and these also can be combined), we need to do a full parsing loop, not just consider the first one.