ENH: Offer inout context or shell to investigate when deciding to fast-forward to tracked remote
Deciding this can be tricky; when repeating a prior merge (e.g. after "git unmerge", amending, and "git-reintegratetom"), the commits on the remote tracked branch must _not_ be fast-forwarded, because they are the outdated duplicates of the redone branch. But when reintegrating into an updated upstream, we do want the fast-forward.
Take the context implementation from git-remote-mutate and adapt it.