git-reintegrate: ENH: Detect redoing of a previously merged and pushed branch and change wording of fast-forward query
I often push a just-merged feature, then unmerge to amend or add commit(s), then reintegrate. The question to fast-forward is misleading then, because this is a redo, and a confirmation would have fatal consequences for the merge.
    $ git feat finish
    $ git opush
    $ git unmerge
    $ git amend
    $ git reintegratetom
    The 3 commits on the branch can be fast-forwarded instead of merged. Fast-forward? NO!!!
Let's try to detect this situation through a new git-similarcommits tool that assumes that most commits aren't changed and keep their commit message title.