git-showfiles: FIX: Ranges still not right (after 89fd0d4)
For git log, we need to keep the original range (... is symmetric); this range may or may not be given, so use an array to store it.
If diffRange is given (either via a passed range or if --cached is passed), use git diff, and apply the correction (.. is symmetric).
Fallback (for no intermediate commits and no range) is git-show; that one would take logRange, but as a given range forces use of git diff, it'll always be empty.

(This corrects commit 89fd0d47fa6b89dfbad597de1ef8c05b07e0a277)