ENH: Also handle renames
Only linkify the second (target) filespec, as that's the only one that exists in the commit.
In order to linkify the first (source) filespec, we'd have to memorize the previous commit, but with the default log output order, that will come only afterwards (and we can't even be sure of that because --reverse could have been passed). The same applies to committed file deletions.