FIX: Wrong place for detecting redoing of a previously merged and pushed branch (after 7e058e4)
It's not in the fast-forwarding of the to-be-reintegrated branch that the fast-forwarding may happen, but before that in the forwarding of origin. (That what happens when I develop features from memory, instead of re-creating the actual situation!)
So the message we want to change is not:
> The 3 commits on the branch can be fast-forwarded instead of merged. Fast-forward? NO!!!
But rather:
> Your master branch is not up-to-date with origin, but can be fast-forwarded.

Move the added code up into confirmFastForward().

(This corrects commit 7e058e4d66474ac7ed8e5ea016759a90f1d1c88a)