FIX: br...lcby needs to move the other-command up in the call hierarchy
Because of the --one-more-command; this takes the GIT-COMMAND's arguments and invokes the special command with those. Therefore, those arguments must already be the resolved log arguments, not something that is still wrapped in others-command.
We can now inject the AUTHORS at the right position (i.e. after the effective GIT-COMMAND lc), and use --keep-position on others-command.
The wrapping order now is:
1. any branch-command to process unrelated arguments
2. dashdash-default-command to add the -- if it's not yet there as -- AUTHOR-PATTERN, directly before the others-command invocation
3. others-command to query and resolve
4. rev-range to take the --authors argument and build the range, and then invoke lc (+ log) on it