git-push: ENH: publish tracking doesn't differentiate between already tracked or not
I'm missing the information that tracking would still be added when the remote branch already exists.
Move the re-tracking check up, so that we can leverage the $trackedBranch information. (I think it also semantically makes more sense to first check what we're leaving before looking into where we're going.)
Differentiate between no tracked branch (and then mention that we would establish tracking), and add an info message on the no-op case where we're already tracking the existing branch. (So publish would just be a regular push.)