git-whodid: ENH: Allow per-{dir,file} most-committed log via --each-{file,dir} [--recursive]
Though we could also support the other output variants, I think it's better usability if that -1 option is implied, as the other (longer) outputs make much less sense and therefore are unlikely to be used. If the user really wants something like that, they can easily build their own via an explicit combination of git-eachFile and git-whodid.