CHG: Switch default --each-{file,dir} format from --most-only to --spark, support other formats
The new sparkline provides the most information (more than --most-only) in a still concise one-line way.
Other formats (other than the non-each shortlog, which would be very verbose) can now be combined with --each-{file,dir}.
For the multi-line formats, the header is changed to be on a separate line, so that the columnization works right.
The "Once:" special summary has to use a tab character (which makes no visible difference) for correct columnization, too.