git-who{contributed,owns}: CHG: Switch default --each-{file,dir} format from --most-only to --spark, support other formats
The new sparkline provides the most information (more than --most-only) in a still concise one-line way.
Other formats can now be combined with --each-{file,dir}. The non-each default format can be selected via --full.
For the multi-line formats, the header is changed to be on a separate line, so that the columnization works right.