Add br...whoownsthosechangedfiles variants
As the finer granularity version of br...whodidthosechangedfiles.
Note: br...whoownsthechanges would require a sort of diff with prepended authors, and would show who changed how many lines on that branch.
Simply using whoowns doesn't work, as the list of files is required. git-files-command could do that, but even with a --since argument that's possible for the {td,week,month,...} variants it would cover the whole set of lines in that timeframe, not just the changed lines.