hubcheats-submodules: Sketch usage of a new ffintegrateto[mn] command
Like reintegrate, the common logic, decisions, and procedures of a reverse-merge into a feature branch so that it can then be fast-forwarded to master should be encoded in a command to reduce the mental load and prevent errors.
This documentation is a sketch of how this should work for me.