CHG: Make --auto-revision the default and drop the lastmsg* variants
Starting with the first found revision is a much better DWIM default; the need to fix this to a particular commit is far less frequent. And fixing to the latest commit is a trivial -r HEAD.
This way we can get rid of the (hard to remember, and also poorly documented) difference between msg* and lastmsg* variants.