git-ffintegrateto: ENH: Differentiate between fast-forward of remote target or discrepancy
In git-reintegrate, we also had confirmFastForward() which only looked into the one direction. With git-inout --is-behind, we can distinguish between the two now.
Split off confirmFastForwardTarget(), as the logic is different (no need for the diff with current branch, and the result of the confirmation is reversed in that "Yes" means sticking with the local target branch).