Also ensure that the source branch is not simply ahead of the tracked remote branch before highlighting the similarity
When the local target branch is simply behind both the tracked remote branch and the source branch (e.g. it has been moved back to prior commits after the source branch has branched off), there likely is high similarity, but that's not the kind we're after with this notification. We're after amends, reorderings, and additional commits that have happened on the source branch vs. the tracked branch (and that would indicate that the reintegration must _not_ happen upon the tracked remote). However, if all there is are some commits that are missing on the target branch, but both exist in the same order on remote and on the source branch, that's a simple fast-forwarding then.