hubcheats-submodules: Maintaining merge needs --no-ff, add details about already on master, GitHub action
The ffintegrateto --no-ff still needs to be implemented, as it's the "wrong way around" a merge cannot be easily forced; Git knows that no merge is actually necessary.
I was confused about superproject's master already merged (but not pushed) while the submodules still need reintegrating; however, this has been a deliberate design decision to minimize the transaction time.