git-ffintegrateto: BUG: --no-submodule-update needs to be forwarded to reintegrateto, not used for the first merge
When we're integrating the superproject without pushing, the last step prior to committing the transaction is the fast-forward reintegration of the submodules. That "subdo reintegratetom --ff-only" will fail if the superproject reintegration has already checked out the submodules again, as it needs an active branch.
I don't see a reason why I would need to influence the submodule update of the first merge; that's only prep stuff, anyway.