FIX: br...subdo broken
Unlike subchanges, subdo does not take a range and determines the affected submodules by itself; submodules need to be passed as --for ... ;.
Therefore, the wrapping in files-command of br...whodidthosechangedfiles has to be taken, this time with br...submodules as the SOURCE-COMMAND, and with the added complexity that the submodules cannot be simply appended, but have to be inserted at a particular position in the middle.
To make matters worse, the semicolon confuses the Git alias parser; https://stackoverflow.com/a/9478636/813602 has a rather simple solution for this: Wrap the entire right-hand side in double quotes.