git-ffintegrateto: BUG: The check for rebase is broken
We cannot use ffdryrun here (as in reintegrate), as this is about the current (source) branch ahead of the target [integration source] branch, not the other way around - ffdryrun would attempt to fast-forward the target onto the source, but it will never happen that the target is equal to (and even ahead of) the stuff we want to integrate to target.
Instead, we need (another) git-inout --is-ahead check, and that even already gets us the number of commits we're ahead.