{git,hub}-wrapper: ENH: Handle -c NAME=VALUE config arguments
In order to properly detect a git-ALIAS wrapper, we need to parse off the config arguments. If there's no alias, we can simply reinject the config arg(s) directly after the Git commit itself. For aliases, there's no good way to pass the config values, unfortunately. A command-line argument could interfere with the alias' own arguments (and each alias would have to recognize and parse the -c to avoid an error if that is passed; it could not just be silently ignored). Therefore it's better to pass those via an exported environment variable; with that, (most) aliases can choose to ignore it, and only where it makes sense it needs to be inserted in front of the Git command(s) (unfortunately via eval syntax that makes the whole command unwieldy, or a separate conversion back to an array).