Add git-rev-range-branchcreation and bc... commands
Dunno if those are useful; I initially thought my git-firstcommit would use a different algorithm that could be an alternative to br... and abr..., but in the end (without a passed --base BASE-BRANCH), this also just delegates to brbase.
The interesting thing here is that it stops short of the first commit (and as that goes forward in history, it's hard to obtain - going backwards is easy via COMMIT~1), so let's keep it.