ENH: Only accept other branches from the same remote that the BRANCH is from
So when we're invoked with -b upstream/foo, only consider other upstream branches, not (as is the default) local branches. That kind of scoping makes much more sense.
This becomes void when --include-remote-refs is passed. Also, if --any is passed we don't need to consider this at all, as any ref will do.