DWIM: Base the range on today's date by default
As we're dealing with logs, these typically have a direct relationship to now and likely cover data that is quite recent. In order to support the visual comparison of multiple outputs, having a fixed starting point (even when the last commit is from yesterday or a few days ago) is very helpful.