Need to do a manual submodule wipe after aborted merge
I've wrapped "git merge", but if that fails and drops you back into the shell, any submodule references will still need to be checked out after committing the merge.
Let's mention this command (but only if there actually are submodules).
For failed rebasing, we don't need to do this as there "git rebase --continue" is used, so we invoke our wrapper which does this automatically.