git-logrange-command: CHG: Support --keep-position, -[N] and +[N], no quoted GIT-COMMAND at first position
Both -[N] and +[N] are needed (though I only need to introduce -[N] here for the msg... variants); -[N] now is a more robust version of --keep-position that limits the scope of replacement, whereas +[N] is used for commands that need the arguments appended at the end, even after user-supplied arguments.
What gets passed as log arguments is the remainder, so --keep-position and appending don't make sense with that.