git-rebaselastfixup: BUG: Fixup is always a no-op
Because the range just goes to the fixup commit itself, and doesn't include the original commit that the fixup is meant to augment.
We need to obtain not just the fixup commit, but also the original commit message, and then do a second search (starting from the fixup commit), and make that commit the start commit for the rebase.