git-reintegrate: FIX: GitHub branch deletion on merge only happens if a PR actually existed
To if we've just pushed the branch but didn't use a PR (e.g. for submodules), the branch is still there.
Instead of checking for PR existence, just query the remote whether the branch still exists; this is easier (and more robust, and maybe also a bit faster) that a PR query.