Use sm_path instead of displaypath for submodules
This makes a difference when submodules are not in the superproject's root but in subdirectories, and you're not in the root but a subdirectory; displaypath then is relative to the current directory, whereas sm_path always is the fixed path from the superproject repo root to the submodule.
The user might like the relative location for sub{list,hashes}[r], but as that command also is used in other scripts, better switch there as well.