Robustness: Retry for 20 seconds to give GitHub some time to auto-delete the branch
Checking for an existing pull-request (prior to pushit()) might be faster, but this opens the can of worms that we cannot choose the remote for checking the pull-requests.
I'd rather add a wait loop, so that we're blocking for 20 seconds if there's no pull request and therefore no automatic deletion on merge.