BUG: Cannot pass -b|--branch to br*uncommit-to-branch
Because of the ordering and the selectedcommit-command requirement that the range is passed after GIT-COMMAND (usually last, with --range-is-last, but the timespan commands use two --since DATE args and therefore cannot use that).
For br*move-to-branch we can work around this with +1 <PARAMETER>, but due to the COMMITS inserted by selectedcommit-command, this isn't possible here.
Allow passing the <commit> in an arbitrary position via --from <commit>, and adapt the aliases (which look very wrong with "... --range-is-last -2 uncommit-to-branch -2 COMMITS ..."; likely an editing error) to use that.