Pass FILE args to GIT-COMMAND as well, and introduce special --pass-file-args for wipe and cat that pass just FILE args
Non-logging Git commands do not typically take FILE arguments; these are just passed to limit the scope of which very first commits are considered (without them, it would be the very first commit by the user in the entire repository). But for wipe and cat, we want to pass the FILE arguments along, as these are important for the GIT-COMMAND as well.
Logging Git commands should receive the FILE arguments, too, as they provide scoping there. We need some special handling for the --with-log-args-except-last, as that changes the parsing.