br...uncommit-to-branch: FIX: Cannot pass additional arguments
As --range-is-last always took the very last argument as the range. Some br... variants can do without --range-is-last because they have a parseable --since (e.g. tduncommit-to-branch), but some require it.
Change the git-selectedcommit-command parsing to remove the range off the last -(N) or +(N) argument instead, include the (fourth) RANGE placeholder in the aliases.