hubcheats-submodules: Add check for amends and submodule reintegration
It seems I forgot the amend on my last PR merge. The subdo checks that all affected submodules are on the main branch now (i.e. have been reintegrated) and the dirty check will warn about an incomplete commit or other files that have not been committed in the superproject.
Note: This command-line is cumbersome; I'll soon define an alias for it.