git-commit: Robustness: Disallow amends of empty index
I noticed that amen[d]push will still create a new commit (with just updated date) and push that, thus triggering the CI build. That's unnecessary and not helpful.
"git commit --allow-empty" suggests that Git disallows creating a commit with the same tree, but that does _not_ cover amends. (Apparently, there the commit already exists and the tree difference is to the previous commit.)
I don't have a amen[d] wrapper, but several aliases supply the amen[d] arguments to my commit-wrapper. Therefore, I can prevent this by recognizing the arguments (in the order given by the alias only).