git-rev-range: Indicate empty range with exit status 4
This command is an intermediate plumbing command; it invokes the "real" Git command. Exit status 1 is commonly used to indicate failure, so we should use another number to indicate this special corner case and thereby enabling the user to differentiate between "no range" and "failure when operating on a valid range".