hubcheats-submodules: FIX: Fast-forwarding of submodule branches is not a separate case
For the submodule, no action is then necessary, but in the superproject we still need to include the (formerly branch, now master) commit. Therefore, it's just a decision on whether to rebase a single commit or merge a set of commits; how that new submodule commit came about is not important here.
That may have been different when there was no branch protection, however, and where I think this originated.