git-co[ot]: Use "git checkout --{ours,theirs}" instead of "git checkout-index"
Though checkout-index can process all conflicts with --all, the "git add --update" is problematic because it also adds unconflicted files to the index, which in the case of submodules we do not want, and which generally is unexpected.
Rather, resolve conflicted files ourselves via git-conflicts, in an implementation taked from git-conflictunresolve.