Add dfiles and {,u,o}[mn]files[mn] variants
These are the symmetric variants that cover both changes on the current branch as well as on the reference branch.
So, bmfiles only lists changed files from the current branch to the merge-base with master (asymmetric), but mfiles lists those changes in addition to any changed files done on master since the current branch has branched off (symmetric).