git-whodid: FIX: --names-only --most-only combination doesn't work
First problem is that the filtering happens before the selector, but by then the counts are already gone.
Split into preFilter and postFilter, and execute around selector. This also avoids that we have to duplicate the filtering out of "<none>" that is normally done via grep in the sed call; now both will run. The postFilter also has to account for the (optional) % sigil that is added by countFieldToPercentage.
Secondly, the former clears shortlogOrderingArg as names should be listed alphabetically, but for determining the one with the highest number we still need the sorting.