Add cohselected and coh[1-7g] variants
For checking out hunks from a previous commit, not the index / HEAD as with coh / cosh.
By passing FILE(s) to selectedcommit-command and then (via --pass-file-args, like with git-catselected) also to coh, we'll already limiting the application to those FILE(s), so we don't need separate variants for coh and cosh.