git-rbr: Consistency: ENH: Also support branch filtering by date here
To have a consistent API with git-brc. Unfortunately, we cannot get "git branch" to directly print last commit timestamps, so we have to do a lookup, just like with git-brfirstcommit, but here it's a straightforward "git log1" with the epoch.
I've built a nice augmentLines helper that makes this pretty straightforward :-)