gitconfig: Add *[xqz] count variants
As 1..7 sometimes isn't enough, and I like the shortcut of appending a single character to the command to specify a fixed range. As I already use several suffixes (-i, -I, -r, -v), there's a relative scarcity of good suffixes. Let's try this:
- x for 12 (roman X = 10, but that's too close to 7, so pick "a dozen"
- q for 25 (almost twice 12, and a "quarter of a hundred"
- z for 100 (nice round next power of 10)
I could imagine Z for 1000, but I don't that's needed.