git-ffintegrateto: Renaming: Minor: isMerge flag to isPerformTheMerge
As I was confused whether this is a merge vs. rebase kind of flag, but it's just whether the final merge is actually done or whether it stops short of it (and lets the user do it via the command given).
It would be better to invert the flag's meaning and name it isMergeDryrun, but that doesn't work so well with the variable defaulting.