git-onelinelog-and-decorations: Need to process any tags at the end if there are tags beyond the last commit
Previously, any tagged commit would also be included in the commit list (so commits ⊇ tagged commits), but now there can be tagged commits ∉ commits.
Skip further processing if a commit has been processed, and add a variation of the rule at the end of the input that just prints any tags from the hold space.